Kernel: Migrate checks to use Cpanel::Kernel::Status
Re-use the logic added to cPanel in v66 that reports all of the
information used in the kernel Security Advisor module - and more.

Besides the new error message for handling exceptions, this change should
not alter the behavior.  This change only alters the source of this
information, which lead to a few long conditional statements.