Kernel: Improve RPM update recommendations on KernelCare
If an update is available, we recommend either waiting for KernelCare or
installing the RPM and rebooting.  However, neither of these suggestions
are appropriate for systems with the corresponding KernelCare patch, as
the patch effectively removes the need to update the kernel RPM.  Thus, we
should distinguish between an outdated kernel with an RPM update available
and an up-to-date kernel with an RPM update available.

However, we should continue to present an "info" message about available
RPMs, but with different suggestions, as installing this update still adds
value.  On KernelCare systems, the kernel will boot up whatever kernel is
selected in GRUB and then, during startup, it will patch the kernel to the
latest version.  Thus, there is a brief window during which the unpatched
kernel is running and it is conceivable that a vulnerability could be
exploited on boot, before the patch is added.  Plus, this patching takes
at least some time during the boot.