Kernel: Suggest reboot on mismatched custom kernels
While we cannot make recommendations about what updates are available or
should be installed when running a custom kernel, it is generally safe to
assume that a mismatch between the boot and running kernel versions
probably requires a reboot, regardless of what special kernel build it is.
Thus, add this recommendation to the output for custom kernels.

This makes all three groups: custom, standard, & KernelCare recommend a
reboot on mismatch (with KC alternatively recommending waiting a few
days).  It is the only recommendation shared by all groups.