Kernel: Add reboot notice on unpatched KernelCare systems
If an updated kernel RPM is installed, but KernelCare has not published a
patch yet, then the system is still vulnerable.  In most cases, one can
safely wait for a KernelCare patch.  However, for severe issues, it may be
prudent to reboot as soon as possible, rather than wait.  Toward that end,
add an "info" message while we are still waiting on a KernelCare patch, so
users are aware that the update hasn't taken effect yet.