Don't offer free tier link if default KC patch (only) is detected.
Case CPANEL-17487: If the default KernelCare patch set is detected, we
do not want to offer an automated upgrade to the free tier because
this will currently undo the default patch set in favor of the free.

Also fixed an undefined variable that we discovered during testing.

Failing tests were deleted. Work done for SWAT-780.