Update message to reflect change in boot versus unpatched kernel.
Case SWAT-1076: When running KernelCare, SecurityAdvisor was detecting
a state where the book kernel was updated, but the running kernel was
not matching the boot version. Additionally, the Cpanel module used
for checking the kernel's status and determining if a reboot was
necessary was not taking into account that when KernelCare was
protecting the system, that the comparison should be made between the
boot kernel and the unpatched kernel. In this change to Security
Advisor, the message has been updated to more accurately describe what
is happening and provide a much softer suggestion to the user that
they reboot the machine (because it's not necessary to remain protected).

See SWAT-1187.