Improve ClamAV version checks.
Case CPANEL-20678: The ClamAV version check is now more tolerant to the
version strings being slightly different but still the same. We now
display the versions that were obtained in the advice block.

If we fail to parse the version, we now give an error advice block. This
could happen in some edge cases, like when the freshclam.conf file is
missing.

The landing page to update ClamAV has been change from
'scripts2/getthemes?modules=1' to 'scripts2/manage_plugins'.