Add new check in Security Advisor for Imunify360.
Case LC-10291: Add the module with basic installed/licensed logic
and some preliminary messages. This is enough to get the new item
to show up in the Security Advisor interface.

Case LC-10301: Add the actual license check that reaches out
to verify.cpanel.net to see which products are listed. The exact
identifiers being checked for here may need to be adjusted to match
what gets used on the Store side.