Update license id check based on real data.
Case LC-10309: The values that were here before were placeholders
based on some internal documentation that didn't have everything we
needed. Now that we've seen a sample of the real data, the product
id check has been updated to match reality.

The product id is 'Imunify360' (not all caps).

The package id follows the format [PROVIDER]-IMUNIFY360-[TYPE],
so the check for \bIMUNIFY360\b was already good enough, although
somewhat redundant with the product id check.