Send unique ids from the messages through to the HTML.
Case LC-10322

These unique ids were already supplied as part of the notice
generation, but they were not used on the page. Adding them as unique
ids on the page should help with the creation of automated tests.

The way it's generating HTML from the JavaScript inside of the
template is kind of weird, but it's consistent with the way the
template already works.