LC-10618: Go to front of Imunify360 instead of deep linking into UI.
In the case where there is a problem with the Imunify360 license,
a deep link into the UI will produce a 404 message rather than
explaining what's actually wrong. This type of license issue should
hopefully never happen outside of a testing environment like ours,
but if it does, the user will get a much more meaningful message
by navigating to the front.