Revert of Imunify360 assessor changes that are not v80-compatible.
Case CPANEL-27631: As part of our v82 work, the Imunify360 assessor
was refactored in a way that it began using a couple of the new
v82-only capabilities of Whostmgr::Store.  This was OK at first
because v80 was not on the latest SA commit from master, but once
it was updated, it began failing.

The fix here is just to revert the changes to the Imunify360 assessor.
It would be nice to have this on the new style, but it's not actually
necessary. Since it needs to be compatible with v80, we don't have
that option.

Exceptions to what was reverted:

* Keeping the path fix for the link into Imunify360 so that any license
  problems will be clearly visible.

* Fix some mocking in the unit tests so they don't send a query to the
  Store product list API (which also happened to be failing at least some
  of the time).

Changelog: Fix compatibility issue with Security Advisor assessor.