No longer license check Imunify360
Case BWG-1416

It was possible for false positives to appear in the security advisor
due to people purchasing Imunify360 directly from CloudLinux or another
provider other than cPanel. The license check for Imunify360 will be
removed until SDI has a method in place for us to query the CloudLinux
license servers for the license status.

Changelog: