Update LiteSpeed advise to not be dependent on jailapache setting
Case CPANEL-36090:  Previously, you could dismiss the advise regarding
LiteSpeed by enabling the jailapache setting.  This was incorrect since
LiteSpeed is unaffected by that setting.  Now, the LiteSpeed advise is
processed prior to checking for jailapache in order to ensure that it is
processed regardless of what the jailapache setting is.

Changelog:  Update LiteSpeed advise to not be dependent on jailapache
 setting