Teach Imunify360 check about Whostmgr::Imunify360::needs_kernelcare
Case CPANEL-37744:  In CPANEL-37720, the
Whostmgr::Imunify360::needs_kernelcare was updated to use Cpanel::OS to
determine the OS distro and version.  It now returns appropriately for
the supported OSs.  In order to keep our code as dry as possible, we
should use that method and remove the _needs_kernelcare and _server_type
subs from Imunify360 Assessor in Security Advisor.

Changelog: